Corrección: Alineación del esquema con la DB física, adición de índices y corrección de errores de tipos en actividades

+++ app/src/main/java/com/agrovet/pos/viewmodels/VentaViewModel.java
@@ -20,8 +20,8 @@ public class VentaViewModel extends AndroidViewModel {
         return repository.getAllVentas();
     }
 
-    public void deleteVenta(long id) {
-        repository.deleteById(id);
+    public void deleteVenta(int id) {
+        repository.deleteById((long)id);
     }
     
     public void addVenta(Venta venta) {
